pridetas editinimas ir nuolaidos

+++ database/migrations/2021_08_17_215356_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
             $table->string('model');
             $table->double('price', 6, 2);
             $table->string('photo', 255);
-            $table->integer('discount_price')->nullable();
+            $table->double('discount_price', 6, 2)->nullable();
             $table->boolean('discount_status')->default(false);
             $table->date('discount_end')->nullable();
             $table->timestamps();
